web/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2
django-dotenv==1.4.2
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
djangorestframework
django-filter
markdown
django-allauth
django-rest-auth
qrcode==7.3.1
Pillow==9.1.0